requirements.txt
openai
python-dotenv
azure
requests